bug fix mysql config

+++ spring-security-fundamental2020/MultipleAuthenticationProvider/src/main/java/com/example/multipleauthenticationprovider/repositories/OtpRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OtpRepository extends JpaRepository<Otp, Integer> {
+public interface OtpRepository extends JpaRepository<Otp, Long> {
 
     Optional<Otp> findOtpByUsername(String username);
 }
